Reorganize backend configuration
Default to the Rust backend but favor the C backends if the appropriate
features are enabled.
