+++ Cargo.toml
@@ -22,6 +22,7 @@ members = ['systest']
 
 [dependencies]
 libc = "0.2"
+cfg-if = "0.1"
 miniz-sys = { path = "miniz-sys", version = "0.1.11", optional = true }
 libz-sys = { version = "1.0", optional = true }
 tokio-io = { version = "0.1.11", optional = true }
